+++ flowapi/flowapi/geography.py
@@ -21,7 +21,9 @@ async def get_geography(aggregation_unit):
     )
     #  Get the reply.
     message = await request.socket.recv_json()
-    current_app.logger.debug(f"Got message: {message}")
+    current_app.flowapi_logger.debug(
+        f"Got message: {message}", request_id=request.request_id
+    )
     try:
         status = message["status"]
     except KeyError:
@@ -29,7 +31,7 @@ async def get_geography(aggregation_unit):
             jsonify({"status": "Error", "msg": "Server responded without status"}),
             500,
         )
-    if status == "done":
+    if status == "completed":
         results_streamer = stream_with_context(stream_result_as_json)(
             message["sql"],
             result_name="features",
@@ -37,7 +39,10 @@ async def get_geography(aggregation_unit):
         )
         mimetype = "application/geo+json"
 
-        current_app.logger.debug(f"Returning {aggregation_unit} geography data.")
+        current_app.flowapi_logger.debug(
+            f"Returning {aggregation_unit} geography data.",
+            request_id=request.request_id,
+        )
         return (
             results_streamer,
             200,