+++ flowapi/flowapi/query_endpoints.py
@@ -24,7 +24,9 @@ async def run_query():
 
     #  Get the reply.
     message = await request.socket.recv_json()
-    current_app.logger.debug(f"Received reply {message}")
+    current_app.flowapi_logger.debug(
+        f"Received reply {message}", request_id=request.request_id
+    )
 
     if "id" in message:
         d = {"Location": url_for(f"query.poll_query", query_id=message["id"])}
@@ -43,16 +45,18 @@ async def poll_query(query_id):
     )
     message = await request.socket.recv_json()
 
-    if message["status"] == "done":
+    if message["status"] == "completed":
         return (
             jsonify({}),
             303,
             {"Location": url_for(f"query.get_query", query_id=message["id"])},
         )
-    elif message["status"] == "running":
-        return jsonify({}), 202
+    elif message["status"] in ("executing", "queued"):
+        return jsonify({"status": message["status"]}), 202
+    elif message["status"] in ("errored", "cancelled"):
+        return jsonify({"status": message["status"]}), 500
     else:
-        return jsonify({}), 404
+        return jsonify({"status": message["status"]}), 404
 
 
 @blueprint.route("/get/<query_id>")
@@ -62,7 +66,9 @@ async def get_query(query_id):
         {"request_id": request.request_id, "action": "get_sql", "query_id": query_id}
     )
     message = await request.socket.recv_json()
-    current_app.logger.debug(f"Got message: {message}")
+    current_app.flowapi_logger.debug(
+        f"Got message: {message}", request_id=request.request_id
+    )
     try:
         status = message["status"]
     except KeyError:
@@ -70,13 +76,15 @@ async def get_query(query_id):
             jsonify({"status": "Error", "msg": "Server responded without status"}),
             500,
         )
-    if message["status"] == "done":
+    if message["status"] == "completed":
         results_streamer = stream_with_context(stream_result_as_json)(
             message["sql"], additional_elements={"query_id": query_id}
         )
         mimetype = "application/json"
 
-        current_app.logger.debug(f"Returning result of query {query_id}.")
+        current_app.flowapi_logger.debug(
+            f"Returning result of query {query_id}.", request_id=request.request_id
+        )
         return (
             results_streamer,
             200,
@@ -86,11 +94,11 @@ async def get_query(query_id):
                 "Content-type": mimetype,
             },
         )
-    elif message["status"] == "running":
+    elif message["status"] in ("executing", "queued"):
         return jsonify({}), 202
-    elif message["status"] == "error":
+    elif message["status"] == "errored":
         return jsonify({"status": "Error", "msg": message["error"]}), 403
-    elif status == "awol":
+    elif status in ("awol", "known"):
         return (jsonify({"status": "Error", "msg": message["error"]}), 404)
     else:
         return jsonify({"status": "Error", "msg": f"Unexpected status: {status}"}), 500