+++ flowmachine/flowmachine/core/table.py
@@ -7,14 +7,17 @@
 Simple utility class that represents arbitrary tables in the
 database.
 """
-import logging
+
 from typing import List
 
+from flowmachine.core.query_state import QueryStateMachine
 from .errors import NotConnectedError
 from .query import Query
 from .subset import subset_factory
 
-logger = logging.getLogger("flowmachine").getChild(__name__)
+import structlog
+
+logger = structlog.get_logger(__name__)
 
 
 class Table(Query):
@@ -114,7 +117,12 @@ class Table(Query):
         # Recorded provided columns to ensure that md5 differs with different columns
         self.columns = columns
         super().__init__()
+        # Table is immediately in a 'finished executing' state
+        q_state_machine = QueryStateMachine(self.redis, self.md5)
+        q_state_machine.enqueue()
+        q_state_machine.execute()
         self._db_store_cache_metadata(compute_time=0)
+        q_state_machine.finish()
 
     @property
     def column_names(self) -> List[str]: