+++ flowmachine/flowmachine/features/subscriber/__init__.py
@@ -19,7 +19,7 @@ from .location_visits import LocationVisits
 from .day_trajectories import DayTrajectories
 from .radius_of_gyration import RadiusOfGyration
 from .displacement import Displacement
-from .total_subscriber_events import TotalSubscriberEvents
+from .event_count import EventCount
 from .event_count import EventCount
 from .subscriber_degree import SubscriberDegree
 from .subscriber_location_cluster import subscriber_location_cluster, HartiganCluster
@@ -63,13 +63,17 @@ from .meaningful_locations import (
 )
 
 from .contact_modal_location_distance import ContactModalLocationDistance
+from .interevent_period import IntereventPeriod
+from .topup_amount import TopUpAmount
+from .topup_balance import TopUpBalance
 from .mds_volume import MDSVolume
 from .event_type_proportion import ProportionEventType
+from .per_contact_event_stats import PerContactEventStats
+from .per_location_event_stats import PerLocationEventStats
 
 __all__ = [
     "RadiusOfGyration",
     "NocturnalEvents",
-    "TotalSubscriberEvents",
     "FirstLocation",
     "CallDays",
     "ModalLocation",
@@ -102,6 +106,8 @@ __all__ = [
     "LocationEntropy",
     "ContactEntropy",
     "EventCount",
+    "PerLocationEventStats",
+    "PerContactEventStats",
     "MeaningfulLocations",
     "MeaningfulLocationsAggregate",
     "MeaningfulLocationsOD",
@@ -111,8 +117,14 @@ __all__ = [
     "ContactEntropy",
     "DistanceCounterparts",
 <<<<<<< HEAD
+<<<<<<< HEAD
     "ContactModalLocationDistance",
 =======
+=======
+    "IntereventPeriod",
+    "TopUpAmount",
+    "TopUpBalance",
+>>>>>>> origin/master
     "MDSVolume",
 >>>>>>> origin/master
     "ContactReciprocal",