+++ flowmachine/tests/conftest.py
@@ -148,10 +148,9 @@ def clean_env(monkeypatch):
     monkeypatch.delenv("LOG_LEVEL", raising=False)
     monkeypatch.delenv("WRITE_LOG_FILE", raising=False)
     monkeypatch.delenv("FLOWDB_PORT", raising=False)
-    monkeypatch.delenv("DB_USER", raising=False)
-    monkeypatch.delenv("DB_PW", raising=False)
-    monkeypatch.delenv("DB_HOST", raising=False)
-    monkeypatch.delenv("DB_NAME", raising=False)
+    monkeypatch.delenv("FLOWDB_USER", raising=False)
+    monkeypatch.delenv("FLOWDB_PASS", raising=False)
+    monkeypatch.delenv("FLOWDB_HOST", raising=False)
     monkeypatch.delenv("DB_CONNECTION_POOL_SIZE", raising=False)
     monkeypatch.delenv("DB_CONNECTION_POOL_OVERFLOW", raising=False)
     monkeypatch.delenv("REDIS_HOST", raising=False)
@@ -179,3 +178,44 @@ def get_length(flowmachine_connect):
     yield lambda query: len(
         pd.read_sql_query(query.get_query(), con=flowmachine_connect.engine)
     )
+
+
+class DummyRedis:
+    """
+    Drop-in replacement for redis.
+    """
+
+    def __init__(self):
+        self._store = {}
+
+    def setnx(self, name, val):
+        if name not in self._store:
+            self._store[name] = val.encode()
+
+    def eval(self, script, numkeys, name, event):
+        current_value = self._store[name]
+        try:
+            self._store[name] = self._store[event][current_value]
+            return self._store[name], current_value
+        except KeyError:
+            return current_value, None
+
+    def hset(self, key, current, next):
+        try:
+            self._store[key][current.encode()] = next.encode()
+        except KeyError:
+            self._store[key] = {current.encode(): next.encode()}
+
+    def set(self, key, value):
+        self._store[key] = value.encode()
+
+    def get(self, key):
+        return self._store.get(key, None)
+
+    def keys(self):
+        return sorted(self._store.keys())
+
+
+@pytest.fixture(scope="function")
+def dummy_redis():
+    return DummyRedis()